backend/pyproject.toml
[project]
name = "defect-backend"
version = "0.1.0"
description = "FastAPI service for defect detection"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "python-multipart>=0.0.9",
    "minio>=7.2",
    "pillow>=10.3",
    "torch>=2.3",
    "torchvision>=0.18",
    "ultralytics>=8.2",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "albumentations>=2.0",
    "scipy>=1.10",
]

[dependency-groups]
dev = ["pytest>=8", "httpx>=0.27", "ruff>=0.4"]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = [".", ".."]
testpaths = ["tests"]
